moving IAuthenticator to base, updating central system code in swarm
