+++ Swarm/src/index.ts
@@ -2,10 +2,14 @@
 // Copyright Contributors to the CitrineOS Project
 //
 // SPDX-License-Identifier: Apache 2.0
-
-import { EventGroup, ICache, ICentralSystem, IMessageHandler, IMessageSender, IModule, IModuleApi, SystemConfig } from '@citrineos/base';
+/* eslint-disable @typescript-eslint/prefer-readonly */
+/* eslint-disable @typescript-eslint/indent */
+/* eslint-disable @typescript-eslint/semi */
+/* eslint-disable @typescript-eslint/quotes */
+/* eslint-disable @typescript-eslint/consistent-type-imports */
+import { EventGroup, IAuthenticator, ICache, ICentralSystem, IMessageHandler, IMessageSender, IModule, IModuleApi, INetworkConnection, SystemConfig } from '@citrineos/base';
 import { MonitoringModule, MonitoringModuleApi } from '@citrineos/monitoring';
-import { CentralSystemImpl, initSwagger, MemoryCache, RabbitMqReceiver, RabbitMqSender, RedisCache } from '@citrineos/util';
+import { Authenticator, CentralSystemImpl, initSwagger, MemoryCache, RabbitMqReceiver, RabbitMqSender, RedisCache, WebsocketNetworkConnection } from '@citrineos/util';
 import { JsonSchemaToTsProvider } from '@fastify/type-provider-json-schema-to-ts';
 import Ajv from "ajv";
 import addFormats from "ajv-formats"
@@ -26,6 +30,8 @@ class CitrineOSServer {
      * Fields
      */
     private _config: SystemConfig;
+    private _authenticator: IAuthenticator;
+    private _networkConnection: INetworkConnection;
     private _centralSystem: ICentralSystem;
     private _logger: Logger<ILogObj>;
     private _server: FastifyInstance;
@@ -82,7 +88,11 @@ class CitrineOSServer {
             return this._ajv.compile(schema);
         });
 
-        this._centralSystem = new CentralSystemImpl(this._config, this._cache, undefined, undefined, this._logger, ajv);
+        this._authenticator = new Authenticator(this._cache, this._logger, new sequelize.DeviceModelRepository(config, this._logger));
+
+        this._networkConnection = new WebsocketNetworkConnection(this._config.util.networkConnection.websocketServers, this._cache, this._logger, this._authenticator);
+
+        this._centralSystem = new CentralSystemImpl(this._config, this._cache, this._createSender(), this._createHandler(), this._networkConnection, this._logger, ajv);
 
         process.on('SIGINT', this.shutdown.bind(this));
         process.on('SIGTERM', this.shutdown.bind(this));
